+++ crates/zng-tp-licenses/src/lib.rs
@@ -266,7 +266,7 @@ pub fn parse_cargo_about(json: &str) -> Result<Vec<LicenseUsed>, serde_json::Err
 /// Panics in case of any error.
 #[cfg(feature = "build")]
 pub fn encode_licenses(licenses: &[LicenseUsed]) -> Vec<u8> {
-    deflate::deflate_bytes(&bincode::serialize(licenses).expect("bincode error"))
+    deflate::deflate_bytes(&postcard::to_allocvec(licenses).expect("postard error"))
 }
 
 /// Encode licenses and write to the output file that is included by [`include_bundle!`].
@@ -297,5 +297,5 @@ macro_rules! include_bundle {
 #[cfg(feature = "bundle")]
 pub fn decode_licenses(bin: &[u8]) -> Vec<LicenseUsed> {
     let bin = inflate::inflate_bytes(bin).expect("invalid bundle deflate binary");
-    bincode::deserialize(&bin).expect("invalid bundle bincode binary")
+    postcard::from_bytes(&bin).expect("invalid bundle postard binary")
 }