+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* **Breaking** Upgrade IPC dependencies.
+    - Upgrade `ipc-channel`.
+    - Replace `bincode` with `postcard`.
 * **Breaking** `CursorImg` is now `non_exhaustive`.
 * **Breaking** Upgrade `VideoMode::refresh_rate` to use `Frequency` unit type.
 * **Breaking** Add `use FrequencyUnits as _` in preludes.